conversor_pdf/requirements.txt
asgiref==3.8.1
chardet==5.2.0
defusedxml==0.7.1
Django==5.1
docx2pdf==0.1.8
et-xmlfile==1.1.0
lxml==5.3.0
odfpy==1.4.1
openpyxl==3.1.5
pillow==10.4.0
python-docx==1.1.2
reportlab==4.2.2
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
unoconv==0.9.0
